Use object references for Action states
Actions are able to change the state of the action chain for the
current event, allowing actions to work together without needing to
know how the other works.

In the initial scheme, each action has an incoming state, and returned
and outgoing state. This required the developer of each action to
remember to return the state in all code paths.

This change instead uses the object reference behaviour. As the state
is a dictionary that is passed into the action, any changes in the
function will automatically propagate back to the behaviour, so there
is no need to return the object.