Validate loaded classes against the core interfaces
When loading using `mewbot.loader`, we do not need to require that
all classes are using the registry system, only that they implement
the correct interfaces. This change detatches the loader from the
component module, and also allows future work to use different base
class systems.

This change also means that loaded files are checked that they are
the correct interface based on the `kind:` field value.